+++ packages/cra-template-edu/CHANGELOG.md
@@ -6,6 +6,7 @@
 - [fix] Specify prettier config in package.json field, instead of separate prettier.config.js file
 - [fix] Upgrade EDS from v6 to v7
 - [fix] Upgrade prettier from 2.7.1 to 2.8.0
+- [new] Add eslint-plugin-testing-library
 
 ## 1.1.0 (2022-11-15)